🔨Refactor. Map Features 정리
